refactor: api 데이터 타입 DrinkType에서 관리하기
