+++ src/components/hooks/useAlcoholMyPage.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query'
 import { getMyPageAlcohol } from 'apis'
-import { MyPageData } from 'apis'
+import { MyPageAlcoholType } from 'types/DrinkType'
 
-export function useAlcoholMyPage(): MyPageData {
+export function useAlcoholMyPage(): MyPageAlcoholType {
   const fallback = {}
   const { data = fallback } = useQuery(['mypage'], () => getMyPageAlcohol())
   return data
